pkg: refactor files
