+++ lib/bmultisig-browser.js
@@ -1,0 +1,28 @@
+/*!
+ * bmultisig-browser.js - a bcoin client libraries.
+ * Copyright (c) 2018, The Bcoin Developers (MIT License).
+ * https://github.com/bcoin-org/bmultisig
+ */
+
+'use strict';
+
+/**
+ * A bmultisig "environment" which exposes all
+ * constructors for primitives, msdb, proposaldb.
+ *
+ * @exports bmultisig
+ * @type {Object}
+ */
+
+const bmultisig = exports;
+
+/*
+ * Expose
+ */
+
+bmultisig.Client = require('./client');
+
+// primitives
+bmultisig.Cosigner = require('./primitives/cosigner');
+
+bmultisig.pkg = require('./pkg');
