+++ lib/http.js
@@ -1,13 +1,12 @@
 /*!
  * http.js - http server for bmultisig
  * Copyright (c) 2018, The Bcoin Developers (MIT License).
- * https://github.com/bcoin-org/bcoin
+ * https://github.com/bcoin-org/bmultisig
  */
 
 'use strict';
 
-const assert = require('assert');
-
+const assert = require('bsert');
 const bcoin = require('bcoin');
 const {TX, Script, Address, Network} = bcoin;
 const HDPublicKey = bcoin.hd.HDPublicKey;
@@ -18,8 +17,8 @@ const sha256 = require('bcrypto/lib/sha256');
 const random = require('bcrypto/lib/random');
 const ccmp = require('bcrypto/lib/ccmp');
 const {Server} = require('bweb');
-const Cosigner = require('./cosigner');
-const RouteList = require('./routelist');
+const Cosigner = require('./primitives/cosigner');
+const RouteList = require('./utils/routelist');
 
 /**
  * Multisig HTTP server
