+++ lib/pkg.js
@@ -1,0 +1,21 @@
+/*!
+ * bmultisig.js - package constants
+ * Copyright (c) 2018, The Bcoin Developers (MIT License).
+ * https://github.com/bcoin-org
+ */
+
+'use strict';
+
+/**
+ * Current version string.
+ * @const {String}
+ */
+
+exports.version = 'v0.0.3';
+
+/**
+ * Repository URL.
+ * @const {String}
+ */
+
+exports.url = 'https://github.com/bcoin-org/bmultisig';
