+++ lib/primitives/cosigner.js
@@ -1,7 +1,7 @@
 /*!
  * cosigner.js - Cosigner
  * Copyright (c) 2018, The Bcoin Developers (MIT License).
- * https://github.com/bcoin-org/bcoin
+ * https://github.com/bcoin-org/bmultisig
  */
 
 'use strict';
