+++ lib/proposal.js
@@ -10,7 +10,7 @@ const bcoin = require('bcoin');
 const {Struct} = require('bufio');
 const {common} = bcoin.wallet;
 const {Outpoint, TX} = bcoin;
-const Cosigner = require('./cosigner');
+const Cosigner = require('./primitives/cosigner');
 const layout = require('./layout').proposaldb;
 
 /**
