+++ lib/wallet.js
@@ -20,7 +20,7 @@ const {MTX} = bcoin;
 
 const ProposalDB = require('./proposaldb');
 const MultisigAccount = require('./account');
-const Cosigner = require('./cosigner');
+const Cosigner = require('./primitives/cosigner');
 const layout = require('./layout').msdb;
 
 /**
